% Generated by roxygen2 (4.0.1): do not edit by hand
\name{network}
\alias{network}
\title{Network simulation code}
\usage{
network(n, method, percent = 1, alpha = 0.5, delta = 0.5, Fn = 1,
  f = 1)
}
\arguments{
\item{n}{Number of nodes.}

\item{method}{String, determines how we simulate the network. Either 'erdos', 'waxman' or 'keeling'. No capitals.}

\item{percent}{Determines the percentage of edges we retain. Used in method='erods'.}

\item{alpha}{Ratio of long edges to short edges. Used in method='waxman' or 'keeling'.}

\item{delta}{Probability of an edge between any two nodes. Used in method='waxman' or 'keeling'.}

\item{Fn}{Number of focal points. Used in method='keeling'.}

\item{f}{Proportion points are moved towards closest focal point. Used in method='keeling'.}
}
\value{
nodes nx2 matrix, rows represent the (x,y) coordinate of a node.

edges Matrix, row i represent the existence of an edge between nodes in first two columns. Third column is set to 0 as default for simulating diseases.

A Adjacency matrix.

node_states Vector with each node state: 0=susceptible; 1=infective; 2=recovered.
}
\description{
This code simulates a network using either the Erdos, Waxman or Keeling network method.
}
\note{
9 December 2014
}
\examples{
net <- network(10,'keeling',1,0.5,0.5,4,0.9)
}
\author{
Brock Hermans <brock.hermans@adelaide.edu.au>
}

