% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sir}
\alias{sir}
\title{SIR simulation}
\usage{
sir(n, beta, gamma, status)
}
\arguments{
\item{n}{Number of individuals.}

\item{beta}{Infectivity rate.}

\item{gamma}{Recovery rate.}

\item{status}{1xn vector of each individuals state: 0=susceptible, 1=infective, 2=recovery}
}
\value{
status 1xn vector of each individuals state after the epidemic. Final epidemic size is sum(status==2)
}
\description{
Description
}
\note{
9 December 2014
}
\examples{
# Test that the SIR code gives expected outcomes
beta <- seq(1,7,1)
par(mfrow=c(3,3))
status=rep(0,100); status[1]=1
for (b in beta){
Ne <- c()
for (j in 1:10000){
tmp=sir(100,b,1,status)
Ne <- c(Ne,sum(tmp==2))
}
hist(Ne,main=paste("Histogram for Ne for beta=",b))
}
}
\author{
Brock Hermans <brock.hermans@adelaide.edu.au>
}

