% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{trainSetB}
\alias{trainSetB}
\title{Simulated dataset for Erdos. Method B}
\format{A data frame with 11590 rows and 10 columns}
\description{
A dataset
}
\details{
\itemize{
 \item beta. input infectivity rate
 \item percent. input percentage of all possible egdes in the network
 \item betaAux. estimated value of beta based on a fully connected network
 \item meanCon. mean number of edges for a node
 \item infecCon. number of edge for the original infected node.
 \item Ne. final epidemic size
 \item pairs. number of pairs
 \item triples. number of triples
 \item triangles. number of triangles
 \item R0. number of new infecitons that the original infected node created
 }
}
\keyword{Erdos}
\keyword{dataset;}

