% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.network}
\alias{plot.network}
\title{Plot network objects}
\usage{
\method{plot}{network}(x, size = 4, ...)
}
\arguments{
\item{x}{Network object.}

\item{size}{Integer of size of points in plot.}
}
\value{
plot of network
}
\description{
Takes a network structure and plots the network. Also plots subgraphs if given
}
\note{
29 August 2014
}
\examples{
network <- network(10,0.5)
plot(network)
}
\author{
Brock Hermans <brock.hermans@adelaide.edu.au
}

